requirements.txt
fastapi>=0.98.0
uvicorn>=0.22.0
google-generativeai>=0.4.0
python-multipart>=0.0.6
requests>=2.31.0
jinja2>=3.1.2
pydantic>=2.0.0
python-dotenv>=1.0.0
